travel-agent/utils/scrapers.py
import requests
from bs4 import BeautifulSoup
import time
from urllib.parse import urljoin, urlparse
from typing import List, Dict, Optional

class WebScraper:
    def __init__(self):
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
        })
    
    def scrape_page(self, url: str) -> Optional[BeautifulSoup]:
        """
        Scrape a web page and return BeautifulSoup object
        """
        try:
            response = self.session.get(url)
            response.raise_for_status()
            
            soup = BeautifulSoup(response.content, 'html.parser')
            return soup
        except Exception as e:
            print(f"Error scraping {url}: {str(e)}")
            return None
    
    def extract_flight_data(self, url: str) -> List[Dict]:
        """
        Extract flight data from a travel website
        """
        # This would contain specific logic to extract flight information
        # depending on the website being scraped
        soup = self.scrape_page(url)
        if not soup:
            return []
        
        flights = []
        # Example extraction logic (would be customized per website)
        # flight_elements = soup.find_all('div', class_='flight-result')
        # for element in flight_elements:
        #     flight = {
        #         'airline': element.find('span', class_='airline').text.strip(),
        #         'price': element.find('span', class_='price').text.strip(),
        #         'duration': element.find('span', class_='duration').text.strip(),
        #     }
        #     flights.append(flight)
        
        return flights
    
    def extract_hotel_data(self, url: str) -> List[Dict]:
        """
        Extract hotel data from a travel website
        """
        soup = self.scrape_page(url)
        if not soup:
            return []
        
        hotels = []
        # Example extraction logic (would be customized per website)
        # hotel_elements = soup.find_all('div', class_='hotel-result')
        # for element in hotel_elements:
        #     hotel = {
        #         'name': element.find('h3', class_='hotel-name').text.strip(),
        #         'price': element.find('span', class_='price').text.strip(),
        #         'rating': element.find('span', class_='rating').text.strip(),
        #     }
        #     hotels.append(hotel)
        
        return hotels
    
    def extract_attractions_data(self, url: str) -> List[Dict]:
        """
        Extract tourist attractions data from a travel website
        """
        soup = self.scrape_page(url)
        if not soup:
            return []
        
        attractions = []
        # Example extraction logic
        # attraction_elements = soup.find_all('div', class_='attraction-item')
        # for element in attraction_elements:
        #     attraction = {
        #         'name': element.find('h3', class_='attraction-name').text.strip(),
        #         'description': element.find('p', class_='description').text.strip(),
        #         'rating': element.find('span', class_='rating').text.strip(),
        #     }
        #     attractions.append(attraction)
        
        return attractions
    
    def rate_limit(self, delay: float = 1.0):
        """
        Implement rate limiting to be respectful to websites
        """
        time.sleep(delay)

class FlightScraper(WebScraper):
    def scrape_flight_prices(self, origin: str, destination: str, date: str):
        # Implementation would scrape flight prices from various sites
        pass

class HotelScraper(WebScraper):
    def scrape_hotel_prices(self, destination: str, check_in: str, check_out: str = None):
        # Implementation would scrape hotel prices from various sites
        pass