travel-agent/requirements.txt
streamlit==1.28.0
openai==0.28.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.24.3